requirements.txt
appdirs==1.4.4
attrs==20.2.0
Automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.1
cssselect==1.1.0
decorator==4.4.2
GrabzIt==3.4.5.1
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itemadapter==0.1.0
itemloaders==1.0.3
jmespath==0.10.0
lxml==4.5.2
parsel==1.6.0
Pillow==7.2.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-telegram-bot==12.8
queuelib==1.5.0
Scrapy==2.3.0
service-identity==18.1.0
six==1.15.0
telegram-send==0.25
tornado==6.0.4
Twisted==20.3.0
w3lib==1.22.0
zope.interface==5.1.0
